Updating pooling method
Finds the average number of generations required to see convergence (find the brightest variant). 
You can change the number of wells that you want to include. By altering num_wells, that is how you can find the number of generations that is expected for convergence. (Can also alter number of variants).